move core tests from `model` to `core`
