+++ packages/model/tsconfig.json
@@ -1,7 +1,7 @@
 {
   "compilerOptions": {
     "target": "esnext",
-    "module": "esnext",
+    "module": "commonjs",
     "outDir": "./lib",
     "rootDir": "./src",
     "declaration": true,
